+++ godot/addons/cyclops_level_builder/menu/editor_toolbar.gd
@@ -51,6 +51,7 @@ func _ready():
 	%Menu.add_action_item(ActionIntersectBlock.new(editor_plugin))
 	%Menu.add_action_item(ActionDeleteSelectedBlocks.new(editor_plugin))
 	%Menu.add_action_item(ActionSnapToGrid.new(editor_plugin))
+	%Menu.add_action_item(ActionMergeVerticesCenter.new(editor_plugin))
 	%Menu.add_separator()
 	%Menu.add_action_item(ActionConvertToMesh.new(editor_plugin))
 	%Menu.add_action_item(ActionExportAsGltf.new(editor_plugin))