+++ doc/change_log.md
@@ -13,4 +13,5 @@
 * Adding tool to export current scene as Godot scene with only native Godot objects.
 * Overhaul of the material manager.  Will now automatically track all materials in project and provide filters to let you focus on subsets of materials.
 * Can drag and drop Texture2D resources from the file explorer into the material directory window to automatically create new materials.  Dragging more than one texture will create an animated texture sequence.
+* Adding command to merge vertices at center.